fix options in Apache module template
There is still an issue, with failing spec, where we want to use the
geoip recipe's options with the mod_geoip recipe, without needing to
provide separate config stanzas for each. Essentially this means the
wrong path is generated in this template currently if you have
configured GeoIP license credentials.