Refactor mod_geoip to use an option, fix DB file for GeoLiteCity
This fixes the issue of mod_geoip config not being generated correctly
when the user specifies GeoIP license details, and also uses the right
DB file basename if the user specifies a URL for the GeoLiteCity
database. This isn't very general, but currently the standard GeoLite
country database and GeoLiteCity are the only ones available freely.